Conflicts:
	AliasDomain.cpp
	BuffSock.cpp
	BuffSock.h
	CTRLSvr.cpp
	DNS.cpp
	ExtAliases.cpp
	FINGSvr.cpp
	Hash.cpp
	LMAILSvr.cpp
	MailSvr.cpp
	MainBSD.cpp
	MessQueue.cpp
	MiscUtils.cpp
	MiscUtils.h
	POP3Svr.cpp
	POP3Utils.cpp
	PSYNCSvr.cpp
	ResLocks.cpp
	SMAILSvr.cpp
	SMAILUtils.cpp
	SMTPSvr.cpp
	SSLBind.cpp
	SendMail.cpp
	SvrUtils.cpp
	SysDep.h
	SysDepSolaris.cpp
	TabIndex.cpp
	UsrUtils.cpp
	XMCrypt.cpp
	win32ssl/include/openssl/opensslconf.h

+++ MD5.h
@@ -35,11 +35,11 @@ typedef struct s_md5_ctx {
 } md5_ctx_t;
 
 void md5_init(md5_ctx_t *mdctx);
-void md5_update(md5_ctx_t *mdctx, unsigned char const *data, unsigned int size);
+void md5_update(md5_ctx_t *mdctx, unsigned char const *data, size_t size);
 void md5_final(md5_ctx_t *mdctx);
 
 void md5_hex(unsigned char *src, char *dst);
 void do_md5_file(FILE *file, long start, long bytes, char *hash);
-void do_md5_string(char const *pass, int passlen, char *hash);
+void do_md5_string(char const *pass, size_t passlen, char *hash);
 
 #endif
